pyproject.toml
[project]
name = "voiceprint"
version = "0.1.0"
description = "Phonetic feature extraction service — F0, formants, VOT, syllable rate from voice recordings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "praat-parselmouth>=0.4.5",
    "numpy>=1.26",
    "python-multipart>=0.0.20",
    "pydantic>=2.9",
    "transformers>=4.45",
    "torch>=2.4",
    "torchaudio>=2.4",
    "phonemizer>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: tests that download large models or take significant time (deselect with '-m \"not slow\"')",
]
